[src/sonic-sairedis] Change sai.profile directory
This commit changes the directory of sai.profile to point to the new
directory location for sai.profile /etc/sai.d/sai.profile file.
This is the location of sai.profile dynamic generation or for
copying the file from the hwsku directory.

Unit tested the code on S6100 for dynamic generation of sai.profile file for
both T0 and T1 and the file was created in /etc/sai.d/sai.profile
Similarly tested the sonic binary on S6000 to make sure that the sai.profile is
copied from the /usr/share/sonic/hwsku/sai.profile to /etc/sai.d/sai.profile.

Signed-off-by: Harish Venkatraman <Harish_Venkatraman@dell.com>